refactor dataset video transform and text processing
