+++ examples/opensora_pku/opensora/data/t2v_dataset.py
@@ -1,77 +1,19 @@
 import copy
 import csv
-import html
 import json
 import logging
 import os
 import random
-import re
-import urllib.parse as ul
 from pathlib import Path
 
-import albumentations
-import cv2
-import ftfy
 import numpy as np
-from bs4 import BeautifulSoup
 from decord import VideoReader
 
 import mindspore as ms
 
+from .transforms import create_video_transforms, t5_text_preprocessing
+
 logger = logging.getLogger()
-
-
-def create_video_transforms(h, w, num_frames, interpolation="bicubic", backend="al", disable_flip=True):
-    """
-    pipeline: flip -> resize -> crop
-    h, w : target resize height, weight
-    NOTE: we change interpolation to bicubic for its better precision and used in SD. TODO: check impact on performance
-    """
-    if backend == "al":
-        # expect rgb image in range 0-255, shape (h w c)
-        from albumentations import CenterCrop, HorizontalFlip, SmallestMaxSize
-
-        targets = {"image{}".format(i): "image" for i in range(num_frames)}
-        mapping = {"bilinear": cv2.INTER_LINEAR, "bicubic": cv2.INTER_CUBIC}
-        if disable_flip:
-            # flip is not proper for horizontal motion learning
-            pixel_transforms = albumentations.Compose(
-                [
-                    SmallestMaxSize(max_size=h, interpolation=mapping[interpolation]),
-                    CenterCrop(h, w),
-                ],
-                additional_targets=targets,
-            )
-        else:
-            # originally used in torch ad code, but not good for non-square video data
-            # also conflict the learning of left-right camera movement
-            pixel_transforms = albumentations.Compose(
-                [
-                    HorizontalFlip(p=0.5),
-                    SmallestMaxSize(max_size=h, interpolation=mapping[interpolation]),
-                    CenterCrop(h, w),
-                ],
-                additional_targets=targets,
-            )
-    elif backend == "ms":
-        # TODO: MindData doesn't support batch transform. can NOT make sure all frames are flipped the same
-        from mindspore.dataset import transforms, vision
-        from mindspore.dataset.vision import Inter
-
-        from .transforms import CenterCrop
-
-        mapping = {"bilinear": Inter.BILINEAR, "bicubic": Inter.BICUBIC}
-        pixel_transforms = transforms.Compose(
-            [
-                vision.RandomHorizontalFlip(),
-                vision.Resize(h, interpolation=mapping[interpolation]),
-                CenterCrop(h, w),
-            ]
-        )
-    else:
-        raise NotImplementedError
-
-    return pixel_transforms
 
 
 class TextVideoDataset:
@@ -172,7 +114,10 @@ class TextVideoDataset:
         self.prev_ok_sample = self.get_replace_data(max_attempts)
         self.require_update_prev = False
         self.token_max_length = token_max_length
-        self.use_text_preprocessing = use_text_preprocessing
+        if use_text_preprocessing is None:
+            self.text_preprocessing = lambda x: x.lower().strip()
+        else:
+            self.text_preprocessing = t5_text_preprocessing
 
     def get_img_cap_list(self):
         raise NotImplementedError
@@ -410,134 +355,6 @@ class TextVideoDataset:
             pixel_values = (pixel_values / 127.5 - 1.0).astype(np.float32)
             return_dict = {"video": pixel_values}
             yield video_name, select_video_frames, return_dict
-
-    def text_preprocessing(self, text):
-        if self.use_text_preprocessing:
-            # The exact text cleaning as was in the training stage:
-            text = self.clean_caption(text)
-            text = self.clean_caption(text)
-            return text
-        else:
-            return text.lower().strip()
-
-    @staticmethod
-    def basic_clean(text):
-        text = ftfy.fix_text(text)
-        text = html.unescape(html.unescape(text))
-        return text.strip()
-
-    def clean_caption(self, caption):
-        caption = str(caption)
-        caption = ul.unquote_plus(caption)
-        caption = caption.strip().lower()
-        caption = re.sub("<person>", "person", caption)
-        # urls:
-        caption = re.sub(
-            r"\b((?:https?:(?:\/{1,3}|[a-zA-Z0-9%])|[a-zA-Z0-9.\-]+[.](?:com|co|ru|net|org|edu|gov|it)[\w/-]*\b\/?(?!@)))",  # noqa
-            "",
-            caption,
-        )  # regex for urls
-        caption = re.sub(
-            r"\b((?:www:(?:\/{1,3}|[a-zA-Z0-9%])|[a-zA-Z0-9.\-]+[.](?:com|co|ru|net|org|edu|gov|it)[\w/-]*\b\/?(?!@)))",  # noqa
-            "",
-            caption,
-        )  # regex for urls
-        # html:
-        caption = BeautifulSoup(caption, features="html.parser").text
-
-        # @<nickname>
-        caption = re.sub(r"@[\w\d]+\b", "", caption)
-
-        # 31C0—31EF CJK Strokes
-        # 31F0—31FF Katakana Phonetic Extensions
-        # 3200—32FF Enclosed CJK Letters and Months
-        # 3300—33FF CJK Compatibility
-        # 3400—4DBF CJK Unified Ideographs Extension A
-        # 4DC0—4DFF Yijing Hexagram Symbols
-        # 4E00—9FFF CJK Unified Ideographs
-        caption = re.sub(r"[\u31c0-\u31ef]+", "", caption)
-        caption = re.sub(r"[\u31f0-\u31ff]+", "", caption)
-        caption = re.sub(r"[\u3200-\u32ff]+", "", caption)
-        caption = re.sub(r"[\u3300-\u33ff]+", "", caption)
-        caption = re.sub(r"[\u3400-\u4dbf]+", "", caption)
-        caption = re.sub(r"[\u4dc0-\u4dff]+", "", caption)
-        caption = re.sub(r"[\u4e00-\u9fff]+", "", caption)
-        #######################################################
-
-        # все виды тире / all types of dash --> "-"
-        caption = re.sub(
-            r"[\u002D\u058A\u05BE\u1400\u1806\u2010-\u2015\u2E17\u2E1A\u2E3A\u2E3B\u2E40\u301C\u3030\u30A0\uFE31\uFE32\uFE58\uFE63\uFF0D]+",  # noqa
-            "-",
-            caption,
-        )
-
-        # кавычки к одному стандарту
-        caption = re.sub(r"[`´«»“”¨]", '"', caption)
-        caption = re.sub(r"[‘’]", "'", caption)
-
-        # &quot;
-        caption = re.sub(r"&quot;?", "", caption)
-        # &amp
-        caption = re.sub(r"&amp", "", caption)
-
-        # ip adresses:
-        caption = re.sub(r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}", " ", caption)
-
-        # article ids:
-        caption = re.sub(r"\d:\d\d\s+$", "", caption)
-
-        # \n
-        caption = re.sub(r"\\n", " ", caption)
-
-        # "#123"
-        caption = re.sub(r"#\d{1,3}\b", "", caption)
-        # "#12345.."
-        caption = re.sub(r"#\d{5,}\b", "", caption)
-        # "123456.."
-        caption = re.sub(r"\b\d{6,}\b", "", caption)
-        # filenames:
-        caption = re.sub(r"[\S]+\.(?:png|jpg|jpeg|bmp|webp|eps|pdf|apk|mp4)", "", caption)
-
-        #
-        caption = re.sub(r"[\"\']{2,}", r'"', caption)  # """AUSVERKAUFT"""
-        caption = re.sub(r"[\.]{2,}", r" ", caption)  # """AUSVERKAUFT"""
-
-        caption = re.sub(self.bad_punct_regex, r" ", caption)  # ***AUSVERKAUFT***, #AUSVERKAUFT
-        caption = re.sub(r"\s+\.\s+", r" ", caption)  # " . "
-
-        # this-is-my-cute-cat / this_is_my_cute_cat
-        regex2 = re.compile(r"(?:\-|\_)")
-        if len(re.findall(regex2, caption)) > 3:
-            caption = re.sub(regex2, " ", caption)
-
-        caption = self.basic_clean(caption)
-
-        caption = re.sub(r"\b[a-zA-Z]{1,3}\d{3,15}\b", "", caption)  # jc6640
-        caption = re.sub(r"\b[a-zA-Z]+\d+[a-zA-Z]+\b", "", caption)  # jc6640vc
-        caption = re.sub(r"\b\d+[a-zA-Z]+\d+\b", "", caption)  # 6640vc231
-
-        caption = re.sub(r"(worldwide\s+)?(free\s+)?shipping", "", caption)
-        caption = re.sub(r"(free\s)?download(\sfree)?", "", caption)
-        caption = re.sub(r"\bclick\b\s(?:for|on)\s\w+", "", caption)
-        caption = re.sub(r"\b(?:png|jpg|jpeg|bmp|webp|eps|pdf|apk|mp4)(\simage[s]?)?", "", caption)
-        caption = re.sub(r"\bpage\s+\d+\b", "", caption)
-
-        caption = re.sub(r"\b\d*[a-zA-Z]+\d+[a-zA-Z]+\d+[a-zA-Z\d]*\b", r" ", caption)  # j2d1a2a...
-
-        caption = re.sub(r"\b\d+\.?\d*[xх×]\d+\.?\d*\b", "", caption)
-
-        caption = re.sub(r"\b\s+\:\s+", r": ", caption)
-        caption = re.sub(r"(\D[,\./])\b", r"\1 ", caption)
-        caption = re.sub(r"\s+", " ", caption)
-
-        caption.strip()
-
-        caption = re.sub(r"^[\"\']([\w\W]+)[\"\']$", r"\1", caption)
-        caption = re.sub(r"^[\'\_,\-\:;]", r"", caption)
-        caption = re.sub(r"[\'\_,\-\:\-\+]$", r"", caption)
-        caption = re.sub(r"^\.\S+$", "", caption)
-
-        return caption.strip()
 
 
 def create_dataloader(
